Move `__experimentalUseHasBehaviorsPanel` to hooks.js
