+++ packages/block-editor/src/components/global-styles/index.js
@@ -1,5 +1,6 @@
 export {
 	__experimentalUseGlobalBehaviors,
+	__experimentalUseHasBehaviorsPanel,
 	useGlobalStylesReset,
 	useGlobalSetting,
 	useGlobalStyle,
@@ -24,9 +25,6 @@ export { default as BorderPanel, useHasBorderPanel } from './border-panel';
 export { default as ColorPanel, useHasColorPanel } from './color-panel';
 export { default as EffectsPanel, useHasEffectsPanel } from './effects-panel';
 export { default as FiltersPanel, useHasFiltersPanel } from './filters-panel';
-export {
-	default as __experimentalBehaviorsPanel,
-	__experimentalUseHasBehaviorsPanel,
-} from './behaviors-panel';
+export { default as __experimentalBehaviorsPanel } from './behaviors-panel';
 export { default as AdvancedPanel } from './advanced-panel';
 export { areGlobalStyleConfigsEqual } from './utils';
